Split run_tests.py into fptaylor.py, generate_tests.py, and run_tests.py
