+++ scripts/config.py
@@ -1,0 +1,3 @@
+import os
+
+fptaylor_exe = os.path.join('..', 'fptaylor')
